notes_backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
argon2-cffi>=23.1.0
pyjwt>=2.8.0
cryptography>=41.0.0
numpy>=1.24.0